Update everypay.php
This error indicates that your module needs updating for compatibility with v6.0.
The dbconnect.php file referenced in the error message was depreciated back in v5.2 and finally removed in v6.0. The module will need updating to use init.php instead.

Please update the module and update to latests code.